Simplify the baseline JIT loop hint call site. https://bugs.webkit.org/show_bug.cgi?id=115052.
Reviewed by Geoffrey Garen.

Moved the watchdog timer check after the JIT optimization check. This
ensures that the JIT opimization counter is incremented on every loop
hint even if the watchdog timer fires.

Removed the code that allows the JIT OSR to happen if the watchdog
timer fires but does not result in a termination. It is extremely rare
that the JIT optimization counter would trigger an OSR on the same pass
as when the watchdog timer fire. If it does happen, we'll simply hold
off on servicing the watchdog timer until the next pass (because it's
not time critical).

* jit/JITOpcodes.cpp:
(JSC::JIT::emit_op_loop_hint):
(JSC::JIT::emitSlow_op_loop_hint):



Identifier: 133488@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@148989 268f45cc-cd09-0410-ab3c-d52691b4dbfc