Layer-backed WebViews don't repaint content outside the visible area https://bugs.webkit.org/show_bug.cgi?id=115275
Reviewed by Beth Dakin.

Source/WebCore:

Export ScrollView::setPaintsEntireContents(bool).

* WebCore.exp.in:

Source/WebKit/mac:

When a client of WebView makes the WebView be layer-backed, then
AppKit propagaes the layer-backing down to the WebHTMLView (but
no further).

In this situation, we had a bug where content that was not
in the visible content rect would not get repainted, because
WebCore clips repaints to the visible rect.

Fix by calling ScrollView::setPaintsEntireContents(bool) from
-[WebHTMLView setLayer:], saying that we need
to paint the entire contents if we have a layer.

* WebView/WebHTMLView.mm:
(-[WebHTMLView setLayer:]):

Identifier: 133703@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@149207 268f45cc-cd09-0410-ab3c-d52691b4dbfc