[BlackBerry] Replace disappearing fillBuffer() API with graphics context drawing https://bugs.webkit.org/show_bug.cgi?id=115360 Internal PR 303048.
Patch by Jakob Petsovits <jpetsovits@blackberry.com> on 2013-04-29
Reviewed by Rob Buis.

Instead of using fillBuffer() to draw directly to the
target buffer, we now lock a Drawable on it and fill it
with PlatformGraphicsContext::addPredefinedPattern().

As a bonus, this also includes related clean-ups -
simpler checkerboard painting code, removal of
fillWindow(), clearWindow() and paintDefaultBackground(),
as well as getting rid of the DEBUG_CHECKERBOARD define
which has been useless for performance tracing purposes
for a while now.

* Api/BackingStore.cpp:
(BlackBerry::WebKit::BackingStorePrivate::blitVisibleContents):
* Api/BackingStore_p.h:

Identifier: 133820@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@149336 268f45cc-cd09-0410-ab3c-d52691b4dbfc