[TestWebKitAPI] <WebKit2/WebKit2_C.h> being included in GTK WK1-only builds https://bugs.webkit.org/show_bug.cgi?id=115470
Reviewed by Martin Robinson.

In the TestWebKitAPI's config.h header, only include the <WebKit2/WebKit2_C.h> header when
not building either of the IOS, Windows or GTK-WK1 platforms.

Up until now the header was included for everything but the IOS and Windows platforms due to
an incorrectly stated condition regarding the GTK platform.

* TestWebKitAPI/config.h:


Identifier: 133926@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@149446 268f45cc-cd09-0410-ab3c-d52691b4dbfc