REGRESSION(r140024): child of fixed div inside of absolute div does not get resized when window is resized https://bugs.webkit.org/show_bug.cgi?id=115379
Reviewed by David Hyatt.

Source/WebCore:

When detecting whether a fixed pos object needed to move with an absolute ancestor we checked
whether the logical width of the fixed pos object had changed. This check prevented the fixed pos
object from detecting later on that it needed to relayout its children in layoutBlock(). So recompute the width
for our check instead of updating it.

Test: fast/block/positioning/child-of-fixed-pos-after-movement.html

* rendering/RenderBlock.cpp:
(WebCore::RenderBlock::markFixedPositionObjectForLayoutIfNeeded):

LayoutTests:

* fast/block/positioning/child-of-fixed-pos-after-movement-expected.txt: Added.
* fast/block/positioning/child-of-fixed-pos-after-movement.html: Added.


Identifier: 134103@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@149640 268f45cc-cd09-0410-ab3c-d52691b4dbfc