[Flakiness Dashboard] Remove the Chromium fallback platforms listing from the legend popup https://bugs.webkit.org/show_bug.cgi?id=115636
Reviewed by Ryosuke Niwa.

* TestResultServer/static-dashboards/flakiness_dashboard.js:
(showLegend): Remove the listing of the Chromium fallback platforms from the legend popup. The code iterated
through the platforms in the g_fallbacksMap dictionary to construct this part of the legend, but the dictionary,
while Chromium-specific, can't be removed just yet as there's more code that uses it.


Identifier: 134182@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@149721 268f45cc-cd09-0410-ab3c-d52691b4dbfc