A change in system environment should force all CSS properties to be recomputed. https://bugs.webkit.org/show_bug.cgi?id=115872 <rdar://problem/13781171>
Reviewed by Antti Koivisto.

When a change in the global environment occurs, for example system colors,
we need to make sure all the CSS properties are recomputed and not fetched
from the cache. For this reason we invalidate the properties cache.
I've also renamed the function to give it a more meaningful name.

* page/Page.cpp:
(WebCore::Page::updateStyleForAllPagesAfterGlobalChangeInEnvironment):
* page/Page.h:
* rendering/RenderTheme.cpp:
(WebCore::RenderTheme::platformColorsDidChange):



Identifier: 134299@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@149839 268f45cc-cd09-0410-ab3c-d52691b4dbfc